// HelpDeskFE/HelpDeskFETp/src/api/model/createOrganizationDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateOrganizationDto {
  /**
     * @minLength 0
     * @maxLength 60
     */
  name: string;
  /** @nullable */
  logo?: string | null;
  /**
     * @minLength 0
     * @maxLength 13
     */
  phoneNumber: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minLength 0
     * @maxLength 240
     */
  address: string;
}