// HelpDeskFE/HelpDeskFETp/src/api/model/organizationDtoResponseDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */
import type { OrganizationDto } from './organizationDto';

export interface OrganizationDtoResponseDto {
  data?: OrganizationDto;
  /** @nullable */
  message?: string | null;
  statusCode?: number;
  status?: boolean;
}